+++ tests/testthat/test-group.R
@@ -2,7 +2,7 @@ context('groupOTU')
 
 test_that('groupOTU', {
     nwk <- system.file("extdata", "sample.nwk", package="treeio")
-    tree <- read.tree(nwk)
+    tree <- ape::read.tree(nwk)
     focus <- c("D", "E", "F", "G")
     df <- fortify(groupOTU(tree, focus))
     expect_true(all(df$group[df$label %in% focus] == 1))
@@ -22,7 +22,7 @@ context('groupClade')
 
 test_that('groupClade', {
     nwk <- system.file("extdata", "sample.nwk", package="treeio")
-    tree <- read.tree(nwk)
+    tree <- ape::read.tree(nwk)
     focus <- c("D", "E", "F", "G")
     nodes <- c(21, 17)
     df <- fortify(groupClade(tree, nodes))